2个接口

+++ src/main/java/com/yunhuakeji/attendance/dao/bizdao/model/UserBuildingRef.java
@@ -8,38 +8,26 @@ import java.math.BigDecimal;
 @Table(name = "USER_BUILDING_REF")
 public class UserBuildingRef implements Serializable {
     @Column(name = "USER_ID")
-    private BigDecimal userId;
+    private Long userId;
 
     @Column(name = "BUILDING_ID")
-    private BigDecimal buildingId;
+    private Long buildingId;
 
     private static final long serialVersionUID = 1L;
 
-    /**
-     * @return USER_ID
-     */
-    public BigDecimal getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    /**
-     * @param userId
-     */
-    public void setUserId(BigDecimal userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    /**
-     * @return BUILDING_ID
-     */
-    public BigDecimal getBuildingId() {
+    public Long getBuildingId() {
         return buildingId;
     }
 
-    /**
-     * @param buildingId
-     */
-    public void setBuildingId(BigDecimal buildingId) {
+    public void setBuildingId(Long buildingId) {
         this.buildingId = buildingId;
     }
 }
